Support variable substituion in component call inside component
Move the browser.component action to driver.js to leverage the renderWithVars function defined.
The previous implementation does not work with component_call("$var1", "${var2} suffix") because of missing variable substitution inside browser.component
