+++ templates/components.js
@@ -19,23 +19,4 @@ module.exports.bindComponents = function(browser) {
   }
 <% }); %>
 
-  browser.component = (name, instanceVars) => {
-    browser.perform(() => {
-
-      // get defaults
-      var component = browser.components[name];
-      var defaultsVars = component.defaults;
-      var compVars = Variables.CompVars(browser.vars, defaultsVars, instanceVars)
-
-      // call the component, pushing the new var context onto a stack.
-      browser.compVarStack.push(compVars);
-
-      component.actions();
-
-    })
-
-    return browser;
-
-  }
-
 };
